requirements.txt
fastmcp
httpx
pydantic
cachetools
pyyaml
pyjwt
